requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
feedparser==6.0.11
openai>=1.30.0
apscheduler==3.10.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
